Extract ion mobility array metadata and experiment predicates
