+++ src/openms/source/KERNEL/MSSpectrum.cpp
@@ -13,7 +13,7 @@
 #include <OpenMS/KERNEL/MSSpectrum.h>
 
 #include <OpenMS/FORMAT/PeakTypeEstimator.h>
-#include <OpenMS/IONMOBILITY/IMDataConverter.h>
+#include <OpenMS/IONMOBILITY/IMDataArrayUtils.h>
 
 namespace OpenMS
 {
@@ -809,7 +809,7 @@ namespace OpenMS
   {
     for (index = 0; index < fdas.size(); ++index)
     {
-      if (IMDataConverter::getIMUnit(fdas[index], unit))
+      if (IMDataArrayUtils::getIMUnit(fdas[index], unit))
       {
         return true;
       }
